@@ -1,5 +1,5 @@
 module ActiveRecord
   module SessionStore
-    VERSION = "2.0.0".freeze
+    VERSION = "2.2.0".freeze
   end
 end
